build.go and release.go get compiled together
